refactor(cli): split cli.py into olmlx/cli/ package
Mechanical split of the 3,971-line olmlx/cli.py into per-subcommand
modules: config_cmd, serve, distributed_launch, service, models_cmd,
chat_cmd, bench_cmd, prepare_cmd, parser. The package __init__ re-exports
every name the old module exposed and keeps _COMMAND_HANDLERS +
_resolve_handler (globals()-based dispatch) so
monkeypatch.setattr("olmlx.cli.cmd_*", ...) keeps working.

Rebindable module state (_cli_distributed_*, _atexit_registered,
_signal_handlers_installed) is deliberately not re-exported — a static
re-import would go stale when the owning module rebinds it; app.py now
reads the distributed handoff from olmlx.cli.serve.

Test patch targets whose consumer moved are repointed to the consumer's
new module; stdlib/settings object patches repointed to their canonical
homes (shutil.which, subprocess.run, olmlx.config.settings.*).
