+++ tests/test_shardquant_integration.py
@@ -113,6 +113,7 @@ class TestFindShardDir:
 class TestCliShardPrepare:
     def test_shard_prepare_invokes_calibration(self, tmp_path):
         from olmlx import cli
+        from olmlx.cli import models_cmd, prepare_cmd
 
         args = SimpleNamespace(
             model="org/model",
@@ -126,8 +127,8 @@ class TestCliShardPrepare:
         store.registry.resolve.return_value = None
         store.ensure_downloaded.return_value = tmp_path
         with (
-            patch.object(cli, "_create_store", return_value=store),
-            patch.object(cli, "_configure_logging"),
+            patch.object(models_cmd, "_create_store", return_value=store),
+            patch.object(prepare_cmd, "_configure_logging"),
             patch(
                 "olmlx.engine.shardquant_calibrate.calibrate_model_shard",
                 return_value=tmp_path / "shard",
